+++ scripts/macros/monsterFeatures/soulMonger/soulMonger.js
@@ -1,0 +1,8 @@
+import {soulThirst} from './soulThirst.js';
+import {waveOfWeariness} from './waveOfWeariness.js';
+import {weightOfAges} from './weightOfAges.js';
+export let soulMonger = {
+    'waveOfWeariness': waveOfWeariness,
+    'soulThirst': soulThirst,
+    'weightOfAges': weightOfAges
+}