+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.3.14:
+## New Monster Content:
+- Shadar-kai Soul Monger: Fey Ancestry, Soul Thirst, Weight of Ages, Shadow Dagger, Wave of Weariness
+  
 # 0.3.13:
 ## New Monster Content:
 - Drow House Captain: Hand Crossbow